Player persistence WIP: Now saving highest point and total height.

+++ src/Assets/Scripts/Player/PlayerData.cs
@@ -8,8 +8,8 @@ namespace LegendPeak.Player
     [Serializable]
     public class PlayerData
     {
-        public int highestPoint { get; set; }
-        public int totalHeight { get; set; }
+        public float highestPoint { get; set; }
+        public float totalHeight { get; set; }
         public MonetizedState monetizedState { get; set; }
     }
 }
